pyproject.toml
[tool.poetry]
name = "ai-ticket-creator-backend"
version = "0.1.0"
description = "AI Ticket Creator Backend API for Chrome Extension"
authors = ["AI Ticket Creator <noreply@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "app"}
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
sqlalchemy = ">=2.0.41,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-multipart = ">=0.0.20,<0.0.21"
python-jose = {extras = ["cryptography"], version = ">=3.5.0,<4.0.0"}
passlib = "^1.7.4"
python-dotenv = ">=1.1.1,<2.0.0"
pydantic = ">=2.11.7,<3.0.0"
pydantic-settings = ">=2.10.1,<3.0.0"
alembic = ">=1.16.4,<2.0.0"
pytest = ">=8.4.1,<9.0.0"
pytest-asyncio = ">=1.0.0,<2.0.0"
fastmcp = ">=2.10.5,<3.0.0"
httpx = {extras = ["http2"], version = ">=0.28.1,<0.29.0"}
pyjwt = ">=2.10.1,<3.0.0"
bcrypt = ">=4.3.0,<5.0.0"
websockets = ">=15.0.1,<16.0.0"
openai = ">=1.96.1,<2.0.0"
google-genai = ">=1.25.0,<2.0.0"
pyyaml = ">=6.0.2,<7.0.0"
pandas = ">=2.3.1,<3.0.0"
celery = ">=5.5.3,<6.0.0"
redis = ">=6.2.0,<7.0.0"
flower = ">=2.0.1,<3.0.0"
kombu = ">=5.5.4,<6.0.0"
aiosmtplib = ">=4.0.1,<5.0.0"
jinja2 = ">=3.1.6,<4.0.0"
aiohttp = ">=3.12.14,<4.0.0"
pydantic-ai = ">=0.4.2,<0.5.0"
fastapi-mcp = ">=0.1.0,<1.0.0"
pytz = ">=2025.2,<2026.0"
beautifulsoup4 = "^4.13.4"
pytest-cov = "^6.2.1"
numpy = "^2.3.2"
pillow = "^10.0.0"
puremagic = "^1.26"
pypdf = "^5.1.0"
PyPDF2 = "^3.0.1"
pdfplumber = "^0.11.4"
python-docx = "^1.1.2"
setuptools = "^75.6.0"
marshmallow = ">=3.0.0,<4.0.0"
asyncpg = "^0.30.0"
aiofiles = "^24.1.0"
greenlet = "^3.2.4"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
ruff = "^0.1.0"
pytest-mock = "^3.14.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"